ab_io_type::fixed_capacity_string

Struct FixedCapacityStringU16

pub struct FixedCapacityStringU16<const CAPACITY: usize> { /* private fields */ }

Container for storing a UTF-8 string limited by the specified fixed bytes capacity as u16.

This is a string only by convention, there is no runtime verification done, contents is treated as regular bytes.

See also FixedCapacityStringU8 if you need to store fewer bytes.

This is just a wrapper for FixedCapacityBytesU16 that the type dereferences to with a different semantic meaning.

Implementations

impl<const CAPACITY: usize> FixedCapacityStringU16<CAPACITY>

pub fn try_from_str(s: &str) -> Option<Self>

Try to create an instance from provided string.

Returns None if provided string does not fit into the capacity.

pub fn try_from_bytes(bytes: &[u8]) -> Option<Self>

Try to create an instance from provided bytes.

Returns None if provided bytes do not fit into the capacity.