ab_io_type

Trait IoTypeOptional

pub trait IoTypeOptional: IoType { }

Marker trait, companion to IoType that indicates the ability to store optional contents.

This means that zero bytes size is a valid invariant. This type is never implemented for types implementing TrivialType because they always have fixed size, and it is not zero.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.