ab_io_type::bool

Struct Bool

pub struct Bool { /* private fields */ }

Just like bool, but any bit pattern is valid.

For bool only 0 and 1 are valid bit patterns out of 256 possible, anything else is undefined behavior. This type changes that by treating 0 as false and everything else as true, making it safer to use and allowing it to implement TrivialType.

Implementations

impl Bool

pub const fn new(value: bool) -> Self

Create a new instance from existing boolean value

pub const fn get(&self) -> bool

Get the value

pub const fn set(&mut self, value: bool)

Set new value