ab_io_type::maybe_data

Struct MaybeData

pub struct MaybeData<Data>
where
    Data: TrivialType,
{ /* private fields */ }

Wrapper type for Data that may or may not be filled with contents.

This is somewhat similar to VariableBytes, but instead of variable size, data structure allows either having it or not having the contents, which is a simpler and more convenient API that is also sufficient in many cases.

Implementations

impl<Data> MaybeData<Data>
where Data: TrivialType,

pub const fn from_ref(data: Option<&Data>) -> impl Deref<Target = Self> + '_

Create a new shared instance from provided data reference.

pub fn from_mut<'a>( buffer: &'a mut Data, size: &'a mut u32, ) -> impl DerefMut<Target = Self> + 'a

Create a new exclusive instance from provided data reference.

size can be either 0 or Data::SIZE, indicating that value is missing or present accordingly.

Panics

Panics if size != 0 && size != Data::SIZE

pub fn from_uninit<'a>( uninit: &'a mut MaybeUninit<Data>, size: &'a mut u32, ) -> impl DerefMut<Target = Self> + 'a

Create a new shared instance from provided memory buffer.

size must be 0.

Panics

Panics if size != 0

pub const fn get(&self) -> Option<&Data>

Try to get access to initialized Data, returns None if not initialized

pub fn get_mut(&mut self) -> Option<&mut Data>

Try to get exclusive access to initialized Data, returns None if not initialized

pub fn replace(&mut self, data: Data) -> &mut Data

Initialize by inserting Data by value or replace existing value and return reference to it

pub fn remove(&mut self)

Remove Data inside and turn instance back into uninitialized

pub fn get_mut_or_init_with<Init>(&mut self, init: Init) -> &mut Data
where Init: FnOnce(&mut MaybeUninit<Data>) -> &mut Data,

Get exclusive access to initialized Data, running provided initialization function if necessary

pub unsafe fn assume_init(&mut self) -> &mut Data

Assume value is initialized

Safety

Caller must ensure Data is actually properly initialized

impl<Data> MaybeData<Data>
where Data: TrivialType + Default,

pub fn get_mut_or_default(&mut self) -> &mut Data

Get exclusive access to initialized Data, initializing with default value if necessary