ab_io_type

Trait IoType

pub unsafe trait IoType {
    type PointerType: TrivialType;

    const METADATA: &[u8];

    // Required methods
    fn size(&self) -> u32;
    fn capacity(&self) -> u32;
    unsafe fn set_size(&mut self, size: u32);
    unsafe fn from_ptr<'a>(
        ptr: &'a NonNull<Self::PointerType>,
        size: &'a u32,
        capacity: u32,
    ) -> impl Deref<Target = Self> + 'a;
    unsafe fn from_mut_ptr<'a>(
        ptr: &'a mut NonNull<Self::PointerType>,
        size: &'a mut u32,
        capacity: u32,
    ) -> impl DerefMut<Target = Self> + 'a;
    unsafe fn as_ptr(&self) -> impl Deref<Target = NonNull<Self::PointerType>>;
    unsafe fn as_mut_ptr(
        &mut self,
    ) -> impl DerefMut<Target = NonNull<Self::PointerType>>;
}

Trait that is used for types that are crossing the host/guest boundary in contracts.

Crucially, it is implemented for any type that implements TrivialType and for VariableBytes.

Safety

This trait is used for types with memory transmutation capabilities, it must not be relied on with untrusted data. Serializing and deserializing of types that implement this trait is simply casting of underlying memory. As a result, all the types implementing this trait must not use implicit padding, unions, or anything similar that might make it unsound to access any bits of the type.

Helper functions are provided to make casting to/from bytes a bit safer than it would otherwise, but extra care is still needed.

Do not implement this trait explicitly! Use #[derive(TrivialType)] instead, which will ensure safety requirements are upheld, or use VariableBytes or other provided wrapper types if more flexibility is needed.

In case of variable state size is needed, create a wrapper struct around VariableBytes and implement traits on it by forwarding everything to the inner implementation.

Required Associated Constants

const METADATA: &[u8]

Data structure metadata in binary form, describing shape and types of the contents, see IoTypeMetadataKind for encoding details

Required Associated Types

type PointerType: TrivialType

Pointer with a trivial type that this IoType represents

Required Methods

fn size(&self) -> u32

Number of bytes that are currently used to store data

fn capacity(&self) -> u32

Number of bytes are allocated right now

unsafe fn set_size(&mut self, size: u32)

Set the number of used bytes

Safety

size must be set to number of properly initialized bytes

unsafe fn from_ptr<'a>( ptr: &'a NonNull<Self::PointerType>, size: &'a u32, capacity: u32, ) -> impl Deref<Target = Self> + 'a

Create a reference to a type, which is represented by provided memory.

Memory must be correctly aligned and sufficient in size, but padding beyond the size of the type is allowed. Memory behind a pointer must not be written to in the meantime either.

Only size bytes are guaranteed to be allocated for types that can store a variable amount of data due to the read-only nature of read-only access here.

Safety

Input bytes must be previously produced by taking underlying bytes of the same type.

unsafe fn from_mut_ptr<'a>( ptr: &'a mut NonNull<Self::PointerType>, size: &'a mut u32, capacity: u32, ) -> impl DerefMut<Target = Self> + 'a

Create a mutable reference to a type, which is represented by provided memory.

Memory must be correctly aligned and sufficient in size, or else None will be returned, but padding beyond the size of the type is allowed. Memory behind a pointer must not be read or written to in the meantime either.

size indicates how many bytes are used within a larger allocation for types that can store a variable amount of data.

Safety

Input bytes must be previously produced by taking underlying bytes of the same type.

unsafe fn as_ptr(&self) -> impl Deref<Target = NonNull<Self::PointerType>>

Get a raw pointer to the underlying data with no checks.

Safety

While calling this function is technically safe, it and allows to ignore many of its invariants, so requires extra care. In particular, no modifications must be done to the value while this returned pointer might be used and no changes must be done through the returned pointer. Also, lifetimes are only superficial here and can be easily (and incorrectly) ignored by using Copy.

unsafe fn as_mut_ptr( &mut self, ) -> impl DerefMut<Target = NonNull<Self::PointerType>>

Get an exclusive raw pointer to the underlying data with no checks.

Safety

While calling this function is technically safe, it and allows to ignore many of its invariants, so requires extra care. In particular, the value’s contents must not be read or written to while returned point might be used. Also, lifetimes are only superficial here and can be easily (and incorrectly) ignored by using Copy.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.